+++ chat-widget/src/common/telemetry/TelemetryConstants.ts
@@ -88,8 +88,6 @@ export enum TelemetryEvent {
     GetConversationDetailsCallFailed = "GetConversationDetailsCallFailed",
     EndChatSDKCallFailed = "EndChatSDKCallFailed",
     GetChatReconnectContextSDKCallFailed = "GetChatReconnectContextSDKCallFailed",
-    PostChatContextCallSucceed = "PostChatContextCallSucceed",
-    PostChatContextCallFailed = "PostChatContextCallFailed",
     ParseAdaptiveCardFailed = "ParseAdaptiveCardFailed",
     ClientDataStoreProviderFailed = "ClientDataStoreProviderFailed",
     InMemoryDataStoreFailed = "InMemoryDataStoreFailed",
@@ -101,6 +99,7 @@ export enum TelemetryEvent {
     WidgetLoadComplete = "WidgetLoadComplete",
     WidgetLoadFailed = "WidgetLoadFailed",
     StartChatMethodException = "StartChatMethodException",
+    CloseChatCall = "CloseChatCall",
     CloseChatMethodException = "CloseChatMethodException",
     PrechatSurveyLoaded = "PrechatSurveyLoaded",
     PrechatSubmitted = "PrechatSubmitted",
@@ -123,8 +122,6 @@ export enum TelemetryEvent {
     LoadingPaneLoaded = "LoadingPaneLoaded",
     EmailTranscriptLoaded = "EmailTranscriptLoaded",
     OutOfOfficePaneLoaded = "OutOfOfficePaneLoaded",
-    PostChatSurveyLoadingPaneLoaded = "PostChatSurveyLoadingPaneLoaded",
-    PostChatSurveyLoaded = "PostChatSurveyLoaded",
     ConfirmationPaneLoaded = "ConfirmationPaneLoaded",
     ProactiveChatPaneLoaded = "ProactiveChatPaneLoaded",
     ReconnectChatPaneLoaded = "ReconnectChatPaneLoaded",
@@ -137,6 +134,7 @@ export enum TelemetryEvent {
     SuppressBotMagicCodeSucceeded = "SuppressBotMagicCodeSucceeded",
     SuppressBotMagicCodeFailed = "SuppressBotMagicCodeFailed",
     GetConversationDetailsException = "GetConversationDetailsException",
+    AppStatesException = "AppStatesException",
     BrowserUnloadEventStarted = "BrowserUnloadEventStarted",
     GetAuthTokenCalled = "GetAuthTokenCalled",
     GetAuthTokenFailed = "GetAuthTokenFailed",
@@ -160,6 +158,8 @@ export enum TelemetryEvent {
     DataMaskingRuleApplyFailed = "DataMaskingRuleApplyFailed",
     IC3ClientEvent = "IC3ClientEvent",
     ConversationEndedThreadEventReceived = "ConversationEndedThreadEventReceived",
+    ConversationEndedByCustomer = "ConversationEndedByCustomer",
+    ConversationEndedByAgent = "ConversationEndedByAgent",
     InvalidConfiguration = "InvalidConfiguration",
     SendTypingIndicatorSucceeded = "SendTypingIndicatorSucceeded",
     SendTypingIndicatorFailed = "SendTypingIndicatorFailed",
@@ -186,7 +186,18 @@ export enum TelemetryEvent {
 
     // Internet connection
     NetworkDisconnected = "NetworkDisconnected",
-    NetworkReconnected = "NetworkReconnected"
+    NetworkReconnected = "NetworkReconnected",
+
+    //Post Chat Workflow related
+    LinkModePostChatWorkflowStarted = "LinkModePostChatWorkflowStarted",
+    EmbedModePostChatWorkflowStarted = "EmbedModePostChatWorkflowStarted",
+    PostChatWorkflowFromCustomer = "PostChatWorkflowFromCustomer",
+    PostChatWorkflowFromAgent = "PostChatWorkflowFromAgent",
+    PostChatWorkflowFromBot = "PostChatWorkflowFromBot",
+    PostChatContextCallSucceed = "PostChatContextCallSucceed",
+    PostChatContextCallFailed = "PostChatContextCallFailed",
+    PostChatSurveyLoadingPaneLoaded = "PostChatSurveyLoadingPaneLoaded",
+    PostChatSurveyLoaded = "PostChatSurveyLoaded"
 }
 
 export interface TelemetryInput {
@@ -251,6 +262,11 @@ export class TelemetryConstants {
             case TelemetryEvent.CustomerVoiceResponsePageLoaded:
             case TelemetryEvent.CustomerVoiceFormResponseSubmitted:
             case TelemetryEvent.CustomerVoiceFormResponseError:
+            case TelemetryEvent.LinkModePostChatWorkflowStarted:
+            case TelemetryEvent.EmbedModePostChatWorkflowStarted:
+            case TelemetryEvent.PostChatWorkflowFromCustomer:
+            case TelemetryEvent.PostChatWorkflowFromAgent:
+            case TelemetryEvent.PostChatWorkflowFromBot:
                 return ScenarioType.ACTIONS;
 
             case TelemetryEvent.StartChatSDKCall: