+++ chat-widget/src/components/livechatwidget/common/setPostChatContextAndLoadSurvey.ts
@@ -1,4 +1,4 @@
-import { LogLevel, TelemetryEvent } from "../../../common/telemetry/TelemetryConstants";
+import { BroadcastEvent, LogLevel, TelemetryEvent } from "../../../common/telemetry/TelemetryConstants";
 
 import { BroadcastService } from "@microsoft/omnichannel-chat-components";
 import { ConversationState } from "../../../contexts/common/ConversationState";
@@ -7,9 +7,18 @@ import { ICustomEvent } from "@microsoft/omnichannel-chat-components/lib/types/i
 import { ILiveChatWidgetAction } from "../../../contexts/common/ILiveChatWidgetAction";
 import { LiveChatWidgetActionType } from "../../../contexts/common/LiveChatWidgetActionType";
 import { TelemetryHelper } from "../../../common/telemetry/TelemetryHelper";
+import { ILiveChatWidgetContext } from "../../../contexts/common/ILiveChatWidgetContext";
+import { Constants } from "../../../common/Constants";
+import { endChat } from "./endChat";
+import { ILiveChatWidgetProps } from "../interfaces/ILiveChatWidgetProps";
+import { PostChatSurveyMode } from "../../postchatsurveypanestateful/enums/PostChatSurveyMode";
+import { addDelayInMs } from "../../../common/utils";
+import { NotificationHandler } from "../../webchatcontainerstateful/webchatcontroller/notification/NotificationHandler";
+import { NotificationScenarios } from "../../webchatcontainerstateful/webchatcontroller/enums/NotificationScenarios";
+import { ConversationEndEntity } from "../../../contexts/common/ConversationEndEntity";
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
-export const setPostChatContextAndLoadSurvey = async (chatSDK: any, dispatch: Dispatch<ILiveChatWidgetAction>, persistedChat?: boolean) => {
+const setPostChatContextAndLoadSurvey = async (chatSDK: any, dispatch: Dispatch<ILiveChatWidgetAction>, persistedChat?: boolean) => {
     try {
         if (!persistedChat) {
             // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -34,3 +43,204 @@ export const setPostChatContextAndLoadSurvey = async (chatSDK: any, dispatch: Di
         dispatch({ type: LiveChatWidgetActionType.SET_CONVERSATION_STATE, payload: ConversationState.Postchat });
     });
 };
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+const checkPostChatEnabled = (props: ILiveChatWidgetProps, state: ILiveChatWidgetContext) => {
+    const isPostChatEnabled = props.chatConfig?.LiveWSAndLiveChatEngJoin?.msdyn_postconversationsurveyenable ?? state.domainStates.liveChatConfig?.LiveWSAndLiveChatEngJoin?.msdyn_postconversationsurveyenable;
+    if (isPostChatEnabled !== Constants.true) {
+        return false;
+    }
+    return true;
+};
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+const initiatePostChat = async (props: ILiveChatWidgetProps, chatSDK: any, setAdapter: any, setWebChatStyles: any, dispatch: Dispatch<ILiveChatWidgetAction>, adapter: any, state: ILiveChatWidgetContext) => {
+    const postChatSurveyMode = props.chatConfig?.LiveWSAndLiveChatEngJoin?.msdyn_postconversationsurveymode ?? state.domainStates.liveChatConfig?.LiveWSAndLiveChatEngJoin?.msdyn_postconversationsurveymode;
+
+    // Check if Postchat already in progress
+    if (state.appStates.postChatWorkflowInProgress) {
+        // Handle case where chat is ended by customer
+        if (state.appStates.conversationEndedBy === ConversationEndEntity.Customer) {
+            await endChat(props, chatSDK, setAdapter, setWebChatStyles, dispatch, adapter, false, false, true);
+        }
+    }
+
+    // Conversation Details call required by customer as well as agent 
+    const conversationDetails = await getConversationDetailsCall(chatSDK);
+    // Start Postchat workflow
+    dispatch({ type: LiveChatWidgetActionType.SET_POST_CHAT_WORKFLOW_IN_PROGRESS, payload: true });
+
+    // Below logic checks if chat is ended by customer or agent or bot and handles them separately
+    if (state.appStates.conversationEndedBy === ConversationEndEntity.Customer) {
+        await postChatInitiatedByCustomer(props, chatSDK, setAdapter, setWebChatStyles, dispatch, adapter, state, postChatSurveyMode, conversationDetails);
+    } else if (state.appStates.conversationEndedBy === ConversationEndEntity.Agent) {
+        if (conversationDetails?.participantType === Constants.userParticipantTypeTag) { 
+            await postChatInitiatedByAgent(props, chatSDK, setAdapter, setWebChatStyles, dispatch, adapter, state, postChatSurveyMode);
+        } else if (conversationDetails?.participantType === Constants.botParticipantTypeTag) {
+            await postChatInitiatedByBot(props, chatSDK, setAdapter, setWebChatStyles, dispatch, adapter, state, postChatSurveyMode);
+        } else {
+            TelemetryHelper.logActionEvent(LogLevel.ERROR, {
+                Event: TelemetryEvent.AppStatesException,
+                ExceptionDetails: {
+                    exception: `ConversationDetails was not set correctly: conversationDetails = ${JSON.stringify(conversationDetails)}`
+                }
+            });
+            // Ending chat because something went wrong
+            await endChat(props, chatSDK, setAdapter, setWebChatStyles, dispatch, adapter, false, false, true);
+        }       
+    } else {
+        TelemetryHelper.logActionEvent(LogLevel.ERROR, {
+            Event: TelemetryEvent.AppStatesException,
+            ExceptionDetails: {
+                exception: `Conversation was Ended but App State was not set correctly: conversationEndedBy = ${state.appStates.conversationEndedBy}`
+            }
+        });
+        // Ending chat because something went wrong
+        await endChat(props, chatSDK, setAdapter, setWebChatStyles, dispatch, adapter, false, false, true);
+    }
+};
+
+// Function for link mode postchat workflow which is essentially same for both customer and agent
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+const linkModePostChatWorkflow = (props: any, dispatch: Dispatch<ILiveChatWidgetAction>, setWebChatStyles: any) => {
+    TelemetryHelper.logActionEvent(LogLevel.INFO, {
+        Event: TelemetryEvent.LinkModePostChatWorkflowStarted
+    });
+    dispatch({ type: LiveChatWidgetActionType.SET_CONVERSATION_STATE, payload: ConversationState.InActive });
+
+    // Disable SendBox
+    if (props?.webChatContainerProps?.renderingMiddlewareProps?.hideSendboxOnConversationEnd !== false) {
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+        setWebChatStyles((styles: any) => { return { ...styles, hideSendBox: true }; });
+    }
+};
+
+// Function for embed mode postchat workflow which is essentially same for both customer and agent
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+const embedModePostChatWorkflow = async (props: any, chatSDK: any, setAdapter: any, setWebChatStyles: any, dispatch: Dispatch<ILiveChatWidgetAction>, adapter: any, state: ILiveChatWidgetContext) => {
+    TelemetryHelper.logActionEvent(LogLevel.INFO, {
+        Event: TelemetryEvent.EmbedModePostChatWorkflowStarted
+    });
+    if (state.domainStates.postChatContext) {
+        dispatch({ type: LiveChatWidgetActionType.SET_CONVERSATION_STATE, payload: ConversationState.PostchatLoading });
+        await addDelayInMs(Constants.PostChatLoadingDurationInMs);
+
+        const loadPostChatEvent: ICustomEvent = {
+            eventName: BroadcastEvent.LoadPostChatSurvey,
+        };
+        BroadcastService.postMessage(loadPostChatEvent);
+    } else {
+        TelemetryHelper.logActionEvent(LogLevel.ERROR, {
+            Event: TelemetryEvent.AppStatesException,
+            ExceptionDetails: {
+                exception: `Conversation was Ended but App State was not set correctly: postChatContext = ${state.domainStates.postChatContext}`
+            }
+        });
+        // End chat call since postchat context was not set correctly
+        await endChat(props, chatSDK, setAdapter, setWebChatStyles, dispatch, adapter, false, false, true);
+    }
+};
+
+// Function will handle only postchat cases initiated by customer
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+const postChatInitiatedByCustomer = async (props: any, chatSDK: any, setAdapter: any, setWebChatStyles: any, dispatch: Dispatch<ILiveChatWidgetAction>, adapter: any, state: ILiveChatWidgetContext, postChatSurveyMode: any, conversationDetails: any) => {
+    // Check if agent has joined conversation
+    if (conversationDetails?.canRenderPostChat === Constants.truePascal) {
+        TelemetryHelper.logActionEvent(LogLevel.INFO, {
+            Event: TelemetryEvent.PostChatWorkflowFromCustomer,
+            Description: "PostChat Workflow was started by customer"
+        });
+        const chatSession = await chatSDK?.getCurrentLiveChatContext();
+        if (chatSession) {
+            // eslint-disable-next-line @typescript-eslint/no-explicit-any
+            (chatSDK as any).chatToken = chatSession.chatToken ?? {};
+            chatSDK.requestId = chatSession.requestId;
+        }
+        // End chat call to end chatsdk but not close chat, only if chat ended by customer
+        await endChat(props, chatSDK, setAdapter, setWebChatStyles, dispatch, adapter, false, true, false);
+        if (postChatSurveyMode === PostChatSurveyMode.Embed) {
+            await embedModePostChatWorkflow(props, chatSDK, setAdapter, setWebChatStyles, dispatch, adapter, state);
+        } else if (postChatSurveyMode === PostChatSurveyMode.Link) {
+            linkModePostChatWorkflow(props, dispatch, setWebChatStyles);
+        } else {
+            TelemetryHelper.logActionEvent(LogLevel.ERROR, {
+                Event: TelemetryEvent.AppStatesException,
+                ExceptionDetails: {
+                    exception: `Conversation was Ended but App State was not set correctly: msdyn_postconversationsurveymode = ${postChatSurveyMode}`
+                }
+            });
+            // Ending chat because something went wrong
+            await endChat(props, chatSDK, setAdapter, setWebChatStyles, dispatch, adapter, false, false, true);
+        }
+    } else {
+        // Agent did not join chat so end chat normally
+        await endChat(props, chatSDK, setAdapter, setWebChatStyles, dispatch, adapter, false, false, true);
+    }
+};
+
+// Function will handle only postchat cases initiated by agent
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+const postChatInitiatedByAgent = async (props: any, chatSDK: any, setAdapter: any, setWebChatStyles: any, dispatch: Dispatch<ILiveChatWidgetAction>, adapter: any, state: ILiveChatWidgetContext, postChatSurveyMode: any) => {
+    TelemetryHelper.logActionEvent(LogLevel.INFO, {
+        Event: TelemetryEvent.PostChatWorkflowFromAgent,
+        Description: "PostChat Workflow was started by agent"
+    });
+    if (postChatSurveyMode === PostChatSurveyMode.Embed) {
+        await embedModePostChatWorkflow(props, chatSDK, setAdapter, setWebChatStyles, dispatch, adapter, state);
+    } else if (postChatSurveyMode === PostChatSurveyMode.Link) {
+        linkModePostChatWorkflow(props, dispatch, setWebChatStyles);
+    } else {
+        TelemetryHelper.logActionEvent(LogLevel.ERROR, {
+            Event: TelemetryEvent.AppStatesException,
+            ExceptionDetails: {
+                exception: `Conversation was Ended but App State was not set correctly: msdyn_postconversationsurveymode = ${postChatSurveyMode}`
+            }
+        });
+        // Ending chat because something went wrong
+        await endChat(props, chatSDK, setAdapter, setWebChatStyles, dispatch, adapter, false, false, true);
+    }
+};
+
+// Function will handle only postchat cases initiated by bot
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+const postChatInitiatedByBot = async (props: any, chatSDK: any, setAdapter: any, setWebChatStyles: any, dispatch: Dispatch<ILiveChatWidgetAction>, adapter: any, state: ILiveChatWidgetContext, postChatSurveyMode: any) => {
+    TelemetryHelper.logActionEvent(LogLevel.INFO, {
+        Event: TelemetryEvent.PostChatWorkflowFromBot,
+        Description: "PostChat Workflow was started by bot"
+    });
+    if (postChatSurveyMode === PostChatSurveyMode.Embed) {
+        await embedModePostChatWorkflow(props, chatSDK, setAdapter, setWebChatStyles, dispatch, adapter, state);
+    } else if (postChatSurveyMode === PostChatSurveyMode.Link) {
+        linkModePostChatWorkflow(props, dispatch, setWebChatStyles);
+    } else {
+        TelemetryHelper.logActionEvent(LogLevel.ERROR, {
+            Event: TelemetryEvent.AppStatesException,
+            ExceptionDetails: {
+                exception: `Conversation was Ended but App State was not set correctly: msdyn_postconversationsurveymode = ${postChatSurveyMode}`
+            }
+        });
+        // Ending chat because something went wrong
+        await endChat(props, chatSDK, setAdapter, setWebChatStyles, dispatch, adapter, false, false, true);
+    }
+};
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+const getConversationDetailsCall = async (chatSDK: any) => {
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    let conversationDetails: any = undefined;
+    try {
+        conversationDetails = await chatSDK.getConversationDetails();
+    } catch (error) {
+        TelemetryHelper.logSDKEvent(LogLevel.ERROR, {
+            Event: TelemetryEvent.GetConversationDetailsCallFailed,
+            ExceptionDetails: {
+                exception: `Get Conversation Details Call Failed : ${error}`
+            }
+        });
+        NotificationHandler.notifyError(NotificationScenarios.Connection, "Get Conversation Details Call Failed: " + error);
+    }
+
+    return conversationDetails;
+};
+
+export { setPostChatContextAndLoadSurvey, checkPostChatEnabled, initiatePostChat };