+++ CHANGE_LOG.md
@@ -100,3 +100,5 @@ All notable changes to this project will be documented in this file.## [Unreleas
 - Added PostChat Context check for triggering Embedded PostChat Workflow
 - Fixed PostChat Stories and VRT
 - Adding support for PVA SSO
+- Fixed PostChat Workflows and added telemetry
+- Fixed Postchat loading twice issue